+++ packages/domain-event/CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## 3.3.0
 
-### Minor Changes
+### Patch Changes
 
+- Updated dependencies []:
+  - @volontariapp/messaging@2.5.4
+  - @volontariapp/database@3.2.9
+  - @volontariapp/outbox@0.9.13
 - trigger SQL on event create
 
 ## 3.2.7